pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "jupyqt"
description = "Embed JupyterLab in PySide6 applications"
authors = [{name = "Alexis Jeandet", email = "alexis.jeandet@member.fsf.org"}]
license = "ISC"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.10"
keywords = []
dynamic = ["version"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Framework :: Jupyter",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Typing :: Typed",
]
dependencies = [
    "ipython>=8.14; python_version<'3.14'",
    "ipython>=9.2; python_version>='3.14'",
    "anyio>=4.0",
    "jupyverse[jupyterlab]>=0.14",
    "fps-kernels>=0.11",
    "fps-noauth>=0.10",
    "PySide6>=6.8",
]

[project.urls]
Homepage = "https://SciQLop.github.io/jupyqt"
Documentation = "https://SciQLop.github.io/jupyqt"
Changelog = "https://SciQLop.github.io/jupyqt/changelog"
Repository = "https://github.com/SciQLop/jupyqt"
Issues = "https://github.com/SciQLop/jupyqt/issues"
Discussions = "https://github.com/SciQLop/jupyqt/discussions"
Gitter = "https://gitter.im/jupyqt/community"
Funding = "https://github.com/sponsors/jeandet"

[tool.pdm.version]
source = "call"
getter = "scripts.get_version:get_version"

[tool.pdm.build]
# Include as much as possible in the source distribution, to help redistributors.
excludes = ["**/.pytest_cache"]
source-includes = [
    "config",
    "docs",
    "scripts",
    "share",
    "tests",
    "duties.py",
    "zensical.toml",
    "*.md",
    "LICENSE",
]

[tool.pdm.build.wheel-data]
# Manual pages can be included in the wheel.
# Depending on the installation tool, they will be accessible to users.
# pipx supports it, uv does not yet, see https://github.com/astral-sh/uv/issues/4731.
data = [
    {path = "share/**/*", relative-to = "."},
]

[dependency-groups]
maintain = [
    "git-changelog>=2.5",
    "twine>=5.1",
    "yore>=0.3.3",
]
ci = [
    "duty>=1.6",
    "griffe>=2.0",
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-randomly>=3.15",
    "pytest-xdist>=3.6",
    "pytest-qt>=4.2",
    "ruff>=0.4",
    "ty>=0.0.14",
    "types-markdown>=3.6",
    "types-pyyaml>=6.0",
]
 docs = [
    "ghp-import>=2.1",
    "markdown-callouts>=0.4",
    "markdown-exec>=1.8",
    "mkdocstrings[python]>=0.29",
    # YORE: EOL 3.10: Remove line.
    "tomli>=2.0; python_version < '3.11'",
    "zensical>=0.0.21",
]

[project.entry-points."fps.modules"]
jupyqt_kernel = "jupyqt.server.plugin:JupyQtKernelModule"

[project.entry-points."jupyverse.modules"]
jupyqt_kernel = "jupyqt.server.plugin:JupyQtKernelModule"

[tool.uv]
default-groups = ["maintain", "ci", "docs"]
